requirements.txt
fastapi>=0.112
uvicorn>=0.30
pydantic>=2.7
python-dotenv>=1.0
Jinja2>=3.1
PyYAML>=6.0
pandas>=2.2
numpy>=2.0
# Optional / later:
# ccxt>=4.0
# matplotlib>=3.9
# mplfinance>=0.12
# python-telegram-bot>=21.0
# apscheduler>=3.10